+++ toys-core/lib/toys/core.rb
@@ -30,7 +30,7 @@ module Toys
     # Current version of Toys core.
     # @return [String]
     #
-    VERSION = "0.9.2"
+    VERSION = "0.10.0.dev"
   end
 
   ## @private deprecated