+++ toys-core/lib/toys/middleware.rb
@@ -39,15 +39,15 @@ module Toys
   # middleware, a Toys middleware can wrap execution with its own code,
   # replace it outright, or leave it unmodified.
   #
-  # Generally, a middleware is a class that implements the two methods defined
-  # in this module: {Toys::Middleware#config} and {Toys::Middleware#run}. To
-  # get default implementations that do nothing, a middleware can
-  # `include Toys::Middleware` or subclass {Toys::Middleware::Base}, but this
-  # is not required.
+  # Generally, a middleware is a class that implements one or more of the
+  # methods defined in this module: {Toys::Middleware#config}, and
+  # {Toys::Middleware#run}. This module provides default implementations that
+  # do nothing, but using them is not required. Middleware objects need respond
+  # only to methods they care about.
   #
   module Middleware
     ##
-    # This method is called after a tool has been defined, and gives this
+    # This method is called *after* a tool has been defined, and gives this
     # middleware the opportunity to modify the tool definition. It is passed
     # the tool definition object and the loader, and can make any changes to
     # the tool definition. In most cases, this method should also call
@@ -95,12 +95,6 @@ module Toys
       ##
       # Create a middleware spec.
       #
-      # @overload spec(middleware_object)
-      #   Create a spec wrapping an existing middleware object
-      #
-      #   @param middleware_object [Toys::Middleware] The middleware object
-      #   @return [Toys::Middleware::Spec] A spec
-      #
       # @overload spec(name, *args, **kwargs, &block)
       #   Create a spec indicating a given middleware name should be
       #   instantiated with the given arguments.
@@ -111,27 +105,63 @@ module Toys
       #   @param block [Proc,nil] The block to pass to the constructor
       #   @return [Toys::Middleware::Spec] A spec
       #
+      # @overload spec(array)
+      #   Create a middleware spec from an array specification.
+      #
+      #   The array must be 1-4 elements long. The first element must be the
+      #   middleware name or class. The other three arguments may include any
+      #   or all of the following optional elements, in any order:
+      #    *  An array for the positional arguments to pass to the constructor
+      #    *  A hash for the keyword arguments to pass to the constructor
+      #    *  A proc for the block to pass to the constructor
+      #
+      #   @param array [Array] The array input
+      #   @return [Toys::Middleware::Spec] A spec
+      #
+      # @overload spec(middleware_object)
+      #   Create a spec wrapping an existing middleware object
+      #
+      #   @param middleware_object [Toys::Middleware] The middleware object
+      #   @return [Toys::Middleware::Spec] A spec
+      #
       def spec(middleware, *args, **kwargs, &block)
-        if middleware.is_a?(::String) || middleware.is_a?(::Symbol) || middleware.is_a?(::Class)
+        case middleware
+        when ::Array
+          spec_from_array(middleware)
+        when ::String, ::Symbol, ::Class
           Spec.new(nil, middleware, args, kwargs, block)
+        when Spec
+          middleware
         else
           Spec.new(middleware, nil, nil, nil, nil)
         end
       end
 
       ##
-      # Create a middleware spec from an array specification.
+      # Create a {Toys::Middleware::Stack} from an array of middleware specs.
+      # Each element may be one of the following:
       #
-      # The array must be 1-4 elements long. The first element must be the
-      # middleware name or class. The other three arguments may include any or
-      # all of the following optional elements, in any order:
-      #  *  An array for the positional arguments to pass to the constructor
-      #  *  A hash for the keyword arguments to pass to the constructor
-      #  *  A proc for the block to pass to the constructor
+      #  *  A {Toys::Middleware} object
+      #  *  A {Toys::Middleware::Spec}
+      #  *  An array whose first element is a middleware name or class, and the
+      #     subsequent elements are params that define what to pass to the class
+      #     constructor (see {Toys::Middleware.spec})
       #
-      # @param array [Array] The array input
-      # @return [Toys::Middleware::Spec] A spec
+      # @param input [Array<Toys::Middleware,Toys::Middleware::Spec,Array>]
+      # @return [Toys::Middleware::Stack]
       #
+      def stack(input)
+        case input
+        when Stack
+          input
+        when ::Array
+          Stack.new(default_specs: input.map { |spec| spec(spec) })
+        else
+          raise ::ArgumentError, "Illegal middleware stack: #{input.inspect}"
+        end
+      end
+
+      ## @private
       def spec_from_array(array)
         middleware = array.first
         if !middleware.is_a?(::String) && !middleware.is_a?(::Symbol) && !middleware.is_a?(::Class)
@@ -153,32 +183,6 @@ module Toys
           end
         end
         Spec.new(nil, middleware, args, kwargs, block)
-      end
-
-      ##
-      # Resolve all arguments into an array of middleware specs. Each argument
-      # may be one of the following:
-      #
-      #  *  A {Toys::Middleware} object
-      #  *  A {Toys::Middleware::Spec}
-      #  *  An array whose first element is a middleware name or class, and the
-      #     subsequent elements are params that define what to pass to the class
-      #     constructor (see {Toys::Middleware.spec_from_array})
-      #
-      # @param items [Array<Toys::Middleware,Toys::Middleware::Spec,Array>]
-      # @return [Array<Toys::Middleware::Spec>]
-      #
-      def resolve_specs(*items)
-        items.map do |item|
-          case item
-          when ::Array
-            spec_from_array(item)
-          when Spec
-            item
-          else
-            spec(item)
-          end
-        end
       end
     end
 
@@ -216,14 +220,7 @@ module Toys
         else
           klass = @name
         end
-        # Due to a bug in Ruby < 2.7, passing an empty **kwargs splat to
-        # initialize will fail if there are no formal keyword args.
-        formals = klass.instance_method(:initialize).parameters
-        if @kwargs.empty? && formals.all? { |arg| arg.first != :key && arg.first != :keyrest }
-          klass.new(*@args, &@block)
-        else
-          klass.new(*@args, **@kwargs, &@block)
-        end
+        Compat.instantiate(klass, @args, @kwargs, @block)
       end
 
       ##
@@ -269,6 +266,99 @@ module Toys
         @args = args
         @kwargs = kwargs
         @block = block
+      end
+
+      ## @private
+      def ==(other)
+        other.is_a?(Spec) &&
+          object.eql?(other.object) &&
+          name.eql?(other.name) &&
+          args.eql?(other.args) &&
+          kwargs.eql?(other.kwargs) &&
+          block.eql?(other.block)
+      end
+      alias eql? ==
+
+      ## @private
+      def hash
+        [object, name, args, kwargs, block].hash
+      end
+    end
+
+    ##
+    # A stack of middleware specs.
+    #
+    class Stack
+      ##
+      # The middleware specs that precede the default set.
+      # @return [Array<Toys::Middleware:Spec>]
+      #
+      attr_reader :pre_specs
+
+      ##
+      # The default set of middleware specs.
+      # @return [Array<Toys::Middleware:Spec>]
+      #
+      attr_reader :default_specs
+
+      ##
+      # The middleware specs that follow the default set.
+      # @return [Array<Toys::Middleware:Spec>]
+      #
+      attr_reader :post_specs
+
+      ##
+      # Add a middleware spec to the stack, in the default location, which is
+      # at the end of pre_specs). See {Toys::Middleware.spec} for a description
+      # of the arguments you can pass.
+      #
+      # @overload add(name, *args, **kwargs, &block)
+      # @overload add(array)
+      # @overload add(middleware_object)
+      #
+      def add(middleware, *args, **kwargs, &block)
+        pre_specs.push(Middleware.spec(middleware, *args, **kwargs, &block))
+      end
+
+      ##
+      # Duplicate this stack.
+      #
+      # @return [Toys::Middleware::Stack]
+      #
+      def dup
+        Stack.new(pre_specs: pre_specs.dup,
+                  post_specs: post_specs.dup,
+                  default_specs: default_specs.dup)
+      end
+
+      ##
+      # Build the middleware in this stack.
+      #
+      # @return [Array<Toys::Middleware>]
+      #
+      def build(middleware_lookup)
+        (@pre_specs + @default_specs + @post_specs).map { |spec| spec.build(middleware_lookup) }
+      end
+
+      ## @private
+      def initialize(default_specs: nil, pre_specs: nil, post_specs: nil)
+        @pre_specs = pre_specs || []
+        @post_specs = post_specs || []
+        @default_specs = default_specs || []
+      end
+
+      ## @private
+      def ==(other)
+        other.is_a?(Stack) &&
+          pre_specs.eql?(other.pre_specs) &&
+          default_specs.eql?(other.default_specs) &&
+          post_specs.eql?(other.post_specs)
+      end
+      alias eql? ==
+
+      ## @private
+      def hash
+        [@pre_specs, @default_specs, @post_specs].hash
       end
     end
   end