+++ toys-core/lib/toys/standard_middleware/show_root_version.rb
@@ -28,8 +28,6 @@ module Toys
     # `--version` flag is given.
     #
     class ShowRootVersion
-      include Middleware
-
       ##
       # Default version flags
       # @return [Array<String>]