+++ toys-core/test/lookup-cases/collision/tool-1.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-desc "file tool-1 short description"
-long_desc "file tool-1 long description"
-
-def run
-  puts "file tool-1 execution"
-end